Revert "Explcitly cast to fix format error on i386."
This reverts commit 4d3b156548a9c83e633cabac2021744b9caae684.

+++ sys/vm/vm_map.c
@@ -4111,8 +4111,8 @@ vm_map_printentries(SYSCTL_HANDLER_ARGS)
 	vm_map_lock(map);
 	for (entry = map->header.next, entrynum = 0; entry != &map->header;
 	    entry = entry->next, entrynum++)
-		sbuf_printf(sb, "%u %lu %lu \n", entrynum,
-		    (unsigned long)entry->start, (unsigned long)entry->end);
+		sbuf_printf(sb, "%u %lu %lu \n", entrynum, entry->start,
+		    entry->end);
 	vm_map_unlock(map);
 	sbuf_finish(sb);
 	sbuf_delete(sb);
